+++ CHANGELOG.md
@@ -6,7 +6,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Changed
+
+- Updated `ErrorKind` to match <https://resend.com/docs/api-reference/errors>
+
+### Removed
+
+- `ErrorKind::InvalidToAddress`
+- `ErrorKind::InvalidScope`
 
 ## [0.14.1] - 2025-04-29