+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+## [0.15.0] - 2025-05-09
 
 ### Added
 
@@ -299,6 +301,7 @@ Disabled `reqwest`'s default features and enabled `rustls-tls`.
 
 Initial release.
 
+[0.15.0]: https://crates.io/crates/resend-rs/0.15.0
 [0.14.1]: https://crates.io/crates/resend-rs/0.14.1
 [0.14.0]: https://crates.io/crates/resend-rs/0.14.0
 [0.13.0]: https://crates.io/crates/resend-rs/0.13.0